refactor(#3276): decompose compilePropertyAccess slice 1 — extract 7 guard bands
Wave B (subtask of #3182). Extract 7 cohesive early-return guard bands from
compilePropertyAccess (property-access.ts) into a new sibling module
src/codegen/property-access-dispatch.ts, replacing each with a sentinel-guarded
call (PA_FALLTHROUGH). Pure byte-identical relocation.

Bands: tryDynamicReceiverRuntimeDispatchReads, tryConstructorPrototypeIdentity,
tryPinnedAndDeleteAwareDynamicGet, tryBuiltinNamespaceDeferredReads,
tryBufferViewAttributeReads, tryStandaloneBuiltinAndWasiMemberReads,
tryNativeErrorMemberRead.

property-access.ts: 7989 -> 7107 LOC (-882). New module: 1078 LOC.
Byte-identity: IDENTICAL 39/39 (gc/standalone/wasi). tsc 0. Smoke 7/7.

✓
