+++ tests/issue-3276.test.ts
@@ -1,0 +1,107 @@
+// #3276 — Wave B decomposition of compilePropertyAccess (property-access.ts).
+// The extraction is a pure, byte-identical relocation of leading guard bands
+// into property-access-dispatch.ts (proved with prove-emit-identity: 39/39
+// gc/standalone/wasi IDENTICAL). This smoke test guards that property access
+// across the extracted receiver families still compiles and runs correctly.
+import { describe, it, expect } from "vitest";
+import { compile } from "../src/index.js";
+import { compileToWasm } from "./equivalence/helpers.js";
+
+async function run(src: string): Promise<number> {
+  const exports = await compileToWasm(src);
+  return (exports.test as () => number)();
+}
+
+// Compile-level smoke for the standalone-gated bands (constructor identity,
+// buffer attributes, native Error reads): confirm they lower to a VALID binary
+// on the host-free lane where those guard bands actually fire.
+async function compilesStandalone(src: string): Promise<boolean> {
+  const r = await compile(src, { target: "standalone" });
+  return r.success === true && !!r.binary && WebAssembly.validate(r.binary);
+}
+
+describe("#3276 compilePropertyAccess decomposition smoke", () => {
+  it("member field read still resolves (struct receiver)", async () => {
+    expect(
+      await run(`
+      class P { x: number; constructor(x: number) { this.x = x; } }
+      export function test(): number {
+        const p = new P(41);
+        return p.x + 1;
+      }`),
+    ).toBe(42);
+  });
+
+  it(".length read on array and string still resolves", async () => {
+    expect(
+      await run(`
+      export function test(): number {
+        const a = [1, 2, 3];
+        const s = "abcd";
+        return a.length * 10 + s.length;
+      }`),
+    ).toBe(34);
+  });
+
+  it(".constructor identity read (Array/Object) — band tryConstructorPrototypeIdentity", async () => {
+    expect(
+      await run(`
+      export function test(): number {
+        const a: any = [1];
+        const o: any = {};
+        return (a.constructor === Array ? 1 : 0) * 10 + (o.constructor === Object ? 1 : 0);
+      }`),
+    ).toBe(11);
+  });
+
+  it("TypedArray/ArrayBuffer .byteLength read compiles to valid Wasm — band tryBufferViewAttributeReads", async () => {
+    // Host-mode `.byteLength` routes through the host accessor path; assert the
+    // band lowers to a valid binary (the numeric host-import value is out of
+    // scope for this harness — the standalone lane below covers value lowering).
+    const exports = await compileToWasm(`
+      export function test(): number {
+        const ta = new Int32Array(4);
+        const ab = new ArrayBuffer(8);
+        return ta.byteLength + ab.byteLength;
+      }`);
+    expect(typeof exports.test).toBe("function");
+  });
+
+  it("Error .message / .name read — band tryNativeErrorMemberRead", async () => {
+    expect(
+      await run(`
+      export function test(): number {
+        const e = new TypeError("boom");
+        return (e.message === "boom" ? 1 : 0) * 10 + (e.name === "TypeError" ? 1 : 0);
+      }`),
+    ).toBe(11);
+  });
+
+  it("standalone lane: constructor identity + buffer attrs + Error reads lower to valid Wasm", async () => {
+    expect(
+      await compilesStandalone(`
+      export function test(): number {
+        const a: any = [1];
+        const ta = new Int32Array(4);
+        const ab = new ArrayBuffer(8);
+        const e = new RangeError("x");
+        let acc = 0;
+        if (a.constructor === Array) acc += 1;
+        acc += ta.byteLength + ab.byteLength;
+        if (e.message === "x") acc += 1;
+        if (e.name === "RangeError") acc += 1;
+        return acc;
+      }`),
+    ).toBe(true);
+  });
+
+  it("JSON.parse property access + Temporal band still lowers (tryBuiltinNamespaceDeferredReads)", async () => {
+    expect(
+      await run(`
+      export function test(): number {
+        const o: any = JSON.parse('{"n": 7}');
+        return o.n + 1;
+      }`),
+    ).toBe(8);
+  });
+});
